CURRENT_USER() -> current_user
